Completed implementation
