--- zeg/cloudprovider/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2019 Zegami Ltd
-"""Init."""
